exynos-mem: Fix major security hole
This fixes the exynos-mem device security hole. The driver allowed any user
to access all of the device's lowmem through the provided mmap functionality.

We create a small little framework collecting the actual CMA memory blocks
that exist on the device; they are the root cause of the existence of this device
driver. We white-list only the CMA memory spaces as parameters to the mmap
function and deny access to any other memory space requests.

We furthermore just allow access to the "s3c-fimc" memory block as this is
seemingly the only space which upon access denial actually breaks functionality.

Also inludes commit:

exynos-mem: Add fimc1 to mmap whitelist check

The assumption that only devices which use CONFIG_CMA_DMA use
exynos-mem is wrong. Remove that check.

Newer camera libraries as on Note 2 4.1.2 variants also map fimc directly,
but they are not reserved through kernel CMA allocation in the device file
but through Samsung's S5P framework.

There, the block is called "fimc1", so add that to the whitelist check.

Signed-off-by: Curtis Menard <curtis.menard@gmail.com>